fixed testers
