+++ src/test/java/app/persistence/CustomerRequestMapperTest.java
@@ -1,9 +1,9 @@
-package app.persistence.customer;
+package app.persistence;
 
 import app.entities.Customer;
 import app.entities.CustomerRequest;
 import app.exceptions.DatabaseException;
-import app.persistence.ConnectionPool;
+import app.persistence.customer.CustomerRequestMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -15,30 +15,34 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class CustomerRequestMapperTest {
-/*
+
     private static final String USER = "postgres";
     private static final String PASSWORD = "HigAbt60ig";
     private static final String URL = "jdbc:postgresql://161.35.195.156/%s?currentSchema=public";
     private static final String DB = "carport_test";
-    private static final ConnectionPool connectionPool = ConnectionPool.getInstance(USER, PASSWORD, URL, DB);*/
-
-    private static final String USER ="postgres";
-    private static final String PASSWORD = "postgres";
-    private static final String URL = "jdbc:postgresql://localhost:5432/%s?currentSchema=public";
-    private static final String DB = "carport_test";
     private static final ConnectionPool connectionPool = ConnectionPool.getInstance(USER, PASSWORD, URL, DB);
 
-   @BeforeEach
+    @BeforeEach
     void setup() {
         try (Connection testConnection = connectionPool.getConnection()) {
             try (Statement stmt = testConnection.createStatement()) {
 
-                stmt.execute("DELETE FROM customer_parts_list");
+                stmt.execute("DELETE FROM customer_invoice");
                 stmt.execute("DELETE FROM admin_customer_request");
-                stmt.execute("DELETE FROM customer_request");
+                stmt.execute("DELETE FROM admin_invoice");
+                stmt.execute("DELETE FROM admin_offer");
+                stmt.execute("DELETE FROM admin_parts_list");
                 stmt.execute("DELETE FROM admin");
                 stmt.execute("DELETE FROM customer");
-
+                stmt.execute("DELETE FROM invoice");
+                stmt.execute("DELETE FROM offer");
+                stmt.execute("DELETE FROM parts_list_parts_list_item");
+                stmt.execute("DELETE FROM parts_list");
+                stmt.execute("DELETE FROM parts_list_item");
+                stmt.execute("DELETE FROM material");
+                stmt.execute("DELETE FROM carport");
+                stmt.execute("DELETE FROM price");
+                stmt.execute("DELETE FROM customer_request");
 
                 // Reset sequence number for customer_request_id (assuming it's an auto-incremented sequence)
                 stmt.executeUpdate("ALTER SEQUENCE customer_request_customer_request_id_seq RESTART WITH 1");
@@ -61,8 +65,8 @@ public class CustomerRequestMapperTest {
                 stmt.execute("INSERT INTO customer (first_name, last_name, email, password, phonenumber, address, zip, customer_request_id) VALUES " +
                         "('Morten', 'Hvor blev du af', 'email@email.dk', '1234', 12345678, 'Vejen 2', 2770, 1)");
 
-                stmt.execute("INSERT INTO customer (first_name, last_name, email, password, phonenumber, address, zip, customer_request_id) VALUES " +
-                        "('Lars', 'Larsen', 'email@email.live.dk', '1234', 12345679, 'Vejen 3', 2770, 2)");
+                stmt.execute("INSERT INTO customer (first_name, last_name, email, password, phonenumber, address, zip) VALUES " +
+                        "('Lars', 'Larsen', 'email@email.live.dk', '1234', 12345679, 'Vejen 3', 2770)");
 
 
                 // Set sequence to continue from the largest member_id
@@ -70,7 +74,7 @@ public class CustomerRequestMapperTest {
                 //stmt.execute("SELECT setval('public.customer_customer_id_seq', COALESCE((SELECT MAX(customer_id)+1 FROM public.customer), 1), false))))");
             }
         } catch (SQLException throwables) {
-            fail("Database connection failed"+throwables.getMessage());
+            fail("Database connection failed" + throwables.getMessage());
 
         }
     }
@@ -89,7 +93,7 @@ public class CustomerRequestMapperTest {
             System.out.println(customerRequest.toString());
             System.out.println("--------------------------------");
         }
-        assertEquals(2, customerRequests.size());
+        assertEquals(1, customerRequests.size());
     }
 
     @Test
@@ -98,7 +102,7 @@ public class CustomerRequestMapperTest {
         int firstCustomerRequestId = allRequests.get(0).getCustomerRequestId();
 
         LocalDate localDate = LocalDate.now();
-        CustomerRequest expected = new CustomerRequest(1, 7, 5, 3, "Plasttrapezplader",localDate, "Afventer");
+        CustomerRequest expected = new CustomerRequest(1, 7, 5, 3, "Plasttrapezplader", localDate, "Afventer");
         CustomerRequest actual = new CustomerRequestMapper().getCustomerRequestById(firstCustomerRequestId, connectionPool);
 
         assertEquals(expected.getRequestLength(), actual.getRequestLength());
@@ -110,57 +114,25 @@ public class CustomerRequestMapperTest {
         assertEquals(expected.getStatus(), actual.getStatus());
     }
 
-   @Test
+    @Test
     void deleteCustomerRequest() throws DatabaseException {
         CustomerRequestMapper.deleteCustomerRequest(1, connectionPool);
-        assertEquals(1, new CustomerRequestMapper().getAllCustomerRequest(connectionPool).size());
+        assertEquals(0, new CustomerRequestMapper().getAllCustomerRequest(connectionPool).size());
     }
 
     @Test
     void makeCustomerRequest() throws DatabaseException {
-        Customer currentCustomer = new Customer(0, "test@email.live.dk", "123", 87654321, "Johny", "Deluxe", "test", 2770,"customer");
- 
-        try (Connection connection = connectionPool.getConnection();
-             PreparedStatement insertCustomerStatement = connection.prepareStatement("INSERT INTO customer (email, password, phonenumber, first_name, last_name, address, zip) VALUES (?, ?, ?, ?, ?, ?, ?)", Statement.RETURN_GENERATED_KEYS)) {
+        Customer currentCustomer = new Customer(2, "email@email.live.dk", "1234", 12345679, "Lars", "Larsen", "Vejen 3", 2770, "customer");
+        LocalDate date = LocalDate.now();
+        int height = 300;
+        int width = 400;
+        int length = 600;
+        int expectedCustomerRequestId = 3;
 
-            insertCustomerStatement.setString(1, currentCustomer.getEmail());
-            insertCustomerStatement.setString(2, currentCustomer.getPassword());
-            insertCustomerStatement.setInt(3, currentCustomer.getPhoneNumber());
-            insertCustomerStatement.setString(4, currentCustomer.getFirstName());
-            insertCustomerStatement.setString(5, currentCustomer.getLastName());
-            insertCustomerStatement.setString(6, currentCustomer.getAddress());
-            insertCustomerStatement.setInt(7, currentCustomer.getZip());
+        int customerRequestId = CustomerRequestMapper.makeCustomerRequest(currentCustomer, height, width, length, date, connectionPool);
 
-            int rowsAffected = insertCustomerStatement.executeUpdate();
-
-            if (rowsAffected != 1) {
-                fail("Failed to insert customer");
-            }
-
-            ResultSet generatedKeys = insertCustomerStatement.getGeneratedKeys();
-            if (generatedKeys.next()) {
-                int customerId = generatedKeys.getInt(1);
-
-                LocalDate date = LocalDate.now();
-                int height = 300;
-                int width = 400;
-                int length = 600;
-
-                int initialSize = new CustomerRequestMapper().getAllCustomerRequest(connectionPool).size();
-                CustomerRequestMapper.makeCustomerRequest(currentCustomer, height, width, length, date, connectionPool);
-                int updatedSize = new CustomerRequestMapper().getAllCustomerRequest(connectionPool).size();
-
-                assertEquals(initialSize + 1, updatedSize);
-            } else {
-                fail("Failed to retrieve generated key for customer");
-            }
-        } catch (SQLException e) {
-            fail("Database error: " + e.getMessage());
-        }
+        assertEquals(expectedCustomerRequestId, customerRequestId);
     }
-
-
-
 
 
     @Test
