+++ src/test/java/app/persistence/InvoiceMapperTest.java
@@ -1,9 +1,9 @@
-package app.persistence.customer;
+package app.persistence;
 
 import app.entities.Invoice;
 import app.entities.InvoiceDetails;
 import app.exceptions.DatabaseException;
-import app.persistence.ConnectionPool;
+import app.persistence.customer.InvoiceMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -17,8 +17,8 @@ import static org.junit.jupiter.api.Assertions.*;
 public class InvoiceMapperTest {
 
     private static final String USER = "postgres";
-    private static final String PASSWORD = "postgres";
-    private static final String URL = "jdbc:postgresql://localhost:5432/%s?currentSchema=public";
+    private static final String PASSWORD = "HigAbt60ig";
+    private static final String URL = "jdbc:postgresql://161.35.195.156/%s?currentSchema=public";
     private static final String DB = "carport_test";
     private static final ConnectionPool connectionPool = ConnectionPool.getInstance(USER, PASSWORD, URL, DB);
