+++ src/test/java/app/persistence/PartsListMapperTest.java
@@ -1,61 +1,60 @@
-package app.persistence.admin;
+package app.persistence;
 
 import app.exceptions.DatabaseException;
-import app.persistence.ConnectionPool;
+import app.persistence.admin.PartsListMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.time.LocalDate;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class AdminOfferMapperTest {
+public class PartsListMapperTest {
     private static final String USER = "postgres";
-    private static final String PASSWORD = "postgres";
-    private static final String URL = "jdbc:postgresql://localhost:5432/%s?currentSchema=public";
+    private static final String PASSWORD = "HigAbt60ig";
+    private static final String URL = "jdbc:postgresql://161.35.195.156/%s?currentSchema=public";
     private static final String DB = "carport_test";
     private static final ConnectionPool connectionPool = ConnectionPool.getInstance(USER, PASSWORD, URL, DB);
-
 
     @BeforeEach
     void setUp() {
         try (Connection testConnection = connectionPool.getConnection()) {
             try (Statement stmt = testConnection.createStatement()) {
                 // Remove all rows from relevant tables
+                stmt.execute("DELETE FROM customer");
+                stmt.execute("DELETE FROM public.customer_invoice");
+                stmt.execute("DELETE FROM public.invoice");
                 stmt.execute("DELETE FROM offer");
                 stmt.execute("DELETE FROM parts_list_parts_list_item");
                 stmt.execute("DELETE FROM parts_list");
                 stmt.execute("DELETE FROM parts_list_item");
                 stmt.execute("DELETE FROM material");
+                stmt.execute("DELETE FROM carport");
                 stmt.execute("DELETE FROM price");
+                stmt.execute("DELETE FROM customer_invoice");
                 stmt.execute("DELETE FROM customer_request");
 
                 // Reset the sequence number
-                stmt.execute("SELECT setval('public.offer_offer_id_seq', 1, false)");
                 stmt.execute("SELECT setval('public.parts_list_parts_list_id_seq', 1, false)");
                 stmt.execute("SELECT setval('public.parts_list_item_parts_list_item_id_seq', 1, false)");
                 stmt.execute("SELECT setval('public.material_material_id_seq', 1, false)");
                 stmt.execute("SELECT setval('public.price_price_id_seq', 1, false)");
-                stmt.execute("SELECT setval('public.customer_request_customer_request_id_seq', 1, false)");
 
                 // Insert rows
-                stmt.execute("INSERT INTO customer_request (length, width, height, tile_type, date, status) VALUES " +
-                        "(600, 400, 300, 'Plasttrapezplader', '2024-08-14', 'Klar')");
                 stmt.execute("INSERT INTO material (material_description, height, width, length, price) VALUES " +
                         "('Rem', '45', '195', '480', 20)");
                 stmt.execute("INSERT INTO parts_list_item (material_id, amount, instruction_description, unit, total_price) VALUES " +
                         "(1, 6, 'test', 'Stk.', 100)");
+                stmt.execute("INSERT INTO price (purchase_price, salesprice_with_tax, coverage_ratio) VALUES " +
+                        "(100, 600, 60)");
 
                 // Set sequence to continue from the largest member_id
-                stmt.execute("SELECT setval('public.offer_offer_id_seq', COALESCE((SELECT MAX(offer_id)+1 FROM public.offer), 1), false)");
                 stmt.execute("SELECT setval('public.parts_list_parts_list_id_seq', COALESCE((SELECT MAX(parts_list_id)+1 FROM public.parts_list), 1), false)");
                 stmt.execute("SELECT setval('public.parts_list_item_parts_list_item_id_seq', COALESCE((SELECT MAX(parts_list_item_id)+1 FROM public.parts_list_item), 1), false)");
                 stmt.execute("SELECT setval('public.material_material_id_seq', COALESCE((SELECT MAX(material_id)+1 FROM public.material), 1), false)");
                 stmt.execute("SELECT setval('public.price_price_id_seq', COALESCE((SELECT MAX(price_id)+1 FROM public.price), 1), false)");
-                stmt.execute("SELECT setval('public.customer_request_customer_request_id_seq', COALESCE((SELECT MAX(customer_request_id)+1 FROM public.customer_request), 1), false)");
             }
         } catch (SQLException throwables) {
             fail("Database connection failed");
@@ -68,22 +67,21 @@ public class AdminOfferMapperTest {
     }
 
     @Test
-    void testAddPrice() throws DatabaseException {
-        int expectedPriceId = 1;
+    void testAddPartsList() throws DatabaseException {
+        int expectedPartsListId = 1;
 
-        int actualPriceId = AdminOfferMapper.addPrice(100,300,30,connectionPool);
+        int actualPartsListId = PartsListMapper.addPartsList(1, connectionPool);
 
-        assertEquals(expectedPriceId, actualPriceId);
+        assertEquals(expectedPartsListId, actualPartsListId);
     }
 
     @Test
-    void testAddOffer() throws DatabaseException {
-        int expectedRowsAffected = 1;
-        LocalDate date = LocalDate.of(2024,5,14);
-        int priceId = AdminOfferMapper.addPrice(100,300,30,connectionPool);
-        int partsListId = PartsListMapper.addPartsList(priceId, connectionPool);
+    void testAddPartsListItem() throws DatabaseException {
+        int expectedPartsListId = 1;
+        PartsListMapper.addPartsList(1, connectionPool);
 
-        int actualRowsAffected = AdminOfferMapper.addOffer("test","test","test", date, partsListId, priceId, 1, connectionPool);
-        assertEquals(expectedRowsAffected, actualRowsAffected);
+        int actualPartsListId = PartsListMapper.addPartsListItem(1, 1, connectionPool);
+
+        assertEquals(expectedPartsListId, actualPartsListId);
     }
 }
